requirements.txt
folium==0.16.0
gpxpy==1.6.2
matplotlib==3.8.4
numpy==1.26.4
PyQt5==5.15.10
pyrtlsdr==0.3.0
pyrtlsdrlib==0.0.3
scipy==1.13.0
setuptools==69.5.1
